refactor: представления, связанные с категорией, перенесены в отдельную папку
