+++ Pages/Categories/Delete.cshtml
@@ -1,0 +1,4 @@
+﻿@page "{id:int?}"
+@model RecipeCatalog.Pages.Categories.CategoryDeleteModel
+@{
+}
