+++ Pages/Categories/Edit.cshtml
@@ -1,5 +1,5 @@
 @page
-@model RecipeCatalog.Pages.CategoryModel
+@model RecipeCatalog.Pages.Categories.CategoryModel
 @{
 	if (Model.Category != null && Model.Category.Id != 0) {
 		ViewData["Title"] = Model.Category.Name;
