+++ Pages/Categories/Edit.cshtml.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using RecipeCatalog.Data;
 using RecipeCatalog.Model;
 
-namespace RecipeCatalog.Pages
+namespace RecipeCatalog.Pages.Categories
 {
     public class CategoryModel : PageModel
     {
