+++ Pages/Categories/Index.cshtml
@@ -1,5 +1,5 @@
 @page
-@model RecipeCatalog.Pages.CategoriesModel
+@model RecipeCatalog.Pages.Categories.CategoriesModel
 @{
 	ViewData["Title"] = "Список категорий";
 }
