--- Pages/CategoryDelete.cshtml
@@ -1,4 +1,0 @@
-﻿@page "{id:int?}"
-@model RecipeCatalog.Pages.CategoryDeleteModel
-@{
-}
